requirements.txt
langchain
streamlit
pinecone-client
python-dotenv
PyMuPDF
google-generativeai
google-generativeai langchain-community
